Add adventure and add student
Finished admin for manage adventure and student
