+++ SeaHawk Fitness(?)/SeaHawk Fitness.xcodeproj/xcuserdata/amandaharman.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -24,12 +24,12 @@
             ignoreCount = "0"
             continueAfterRunningActions = "No"
             filePath = "SeaHawk Fitness/DeleteAdventuresViewController.swift"
-            timestampString = "482720613.605749"
+            timestampString = "482877313.134273"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "16"
-            endingLineNumber = "16"
-            landmarkName = "setupCell(_:)"
+            startingLineNumber = "31"
+            endingLineNumber = "31"
+            landmarkName = "viewDidLoad()"
             landmarkType = "5">
          </BreakpointContent>
       </BreakpointProxy>
